pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "hire-comply"
version = "0.1.0"
description = "Compliance-first AI Recruitment Agent powered by LangGraph"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "Haotian"},
]

dependencies = [
    # Core orchestration
    "langgraph>=0.3.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.3.0",
    "langchain-anthropic>=0.3.0",
    "langchain-google-genai>=2.0.0",

    # Backend
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",

    # Database
    "psycopg[binary]>=3.2.0",
    "langgraph-checkpoint-postgres>=2.0.0",

    # Document parsing
    "pypdf>=5.0.0",
    "python-docx>=1.1.0",

    # Utilities
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
    "structlog>=24.0.0",

    # Vector store (optional)
    "chromadb>=0.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "pre-commit>=4.0.0",
]
streamlit = [
    "streamlit>=1.40.0",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
